Update MapsPage component to handle expired tokens
The code changes in the MapsPage component handle expired tokens by decoding the stored token and checking if it has expired. If the token has expired, it is removed from local storage and the user is redirected to the login page. This improves the security and user experience of the application.